moved creatTocItems into helper service
